deps: upgrade npm to 6.10.2

+++ deps/npm/test/common-tap.js
@@ -7,6 +7,12 @@ var readCmdShim = require('read-cmd-shim')
 var isWindows = require('../lib/utils/is-windows.js')
 var Bluebird = require('bluebird')
 
+// remove any git envs so that we don't mess with the main repo
+// when running git subprocesses in tests
+Object.keys(process.env).filter(k => /^GIT/.test(k)).forEach(
+  k => delete process.env[k]
+)
+
 // cheesy hackaround for test deps (read: nock) that rely on setImmediate
 if (!global.setImmediate || !require('timers').setImmediate) {
   require('timers').setImmediate = global.setImmediate = function () {
@@ -16,26 +22,63 @@ if (!global.setImmediate || !require('timers').setImmediate) {
 }
 
 var spawn = require('child_process').spawn
+const spawnSync = require('child_process').spawnSync
 var path = require('path')
+
+// space these out to help prevent collisions
+const testId = 3 * (+process.env.TAP_CHILD_ID || 0)
 
 // provide a working dir unique to each test
 const main = require.main.filename
-exports.pkg = path.resolve(path.dirname(main), path.basename(main, '.js'))
+const testName = path.basename(main, '.js')
+exports.pkg = path.resolve(path.dirname(main), testName)
+var commonCache = path.resolve(__dirname, 'npm_cache_' + testName)
+exports.cache = commonCache
+
 const mkdirp = require('mkdirp')
 const rimraf = require('rimraf')
+rimraf.sync(exports.pkg)
+rimraf.sync(commonCache)
 mkdirp.sync(exports.pkg)
+mkdirp.sync(commonCache)
+// if we're in sudo mode, make sure that the cache is not root-owned
+const isRoot = process.getuid && process.getuid() === 0
+const isSudo = isRoot && process.env.SUDO_UID && process.env.SUDO_GID
+if (isSudo) {
+  const sudoUid = +process.env.SUDO_UID
+  const sudoGid = +process.env.SUDO_GID
+  fs.chownSync(commonCache, sudoUid, sudoGid)
+}
+
+const returnCwd = path.dirname(__dirname)
+const find = require('which').sync('find')
 require('tap').teardown(() => {
+  // work around windows folder locking
+  process.chdir(returnCwd)
   try {
-    rimraf.sync(exports.pkg)
+    if (isSudo) {
+      // running tests as sudo.  ensure we didn't leave any root-owned
+      // files in the cache by mistake.
+      const args = [ commonCache, '-uid', '0' ]
+      const found = spawnSync(find, args)
+      const output = found && found.stdout && found.stdout.toString()
+      if (output.length) {
+        const er = new Error('Root-owned files left in cache!')
+        er.testName = main
+        er.files = output.trim().split('\n')
+        throw er
+      }
+    }
+    if (!process.env.NO_TEST_CLEANUP) {
+      rimraf.sync(exports.pkg)
+      rimraf.sync(commonCache)
+    }
   } catch (e) {
     if (process.platform !== 'win32') {
       throw e
     }
   }
 })
-
-// space these out to help prevent collisions
-const testId = 3 * (+process.env.TAP_CHILD_ID || 0)
 
 var port = exports.port = 15443 + testId
 exports.registry = 'http://localhost:' + port
@@ -53,8 +96,8 @@ ourenv.npm_config_progress = 'false'
 ourenv.npm_config_metrics = 'false'
 ourenv.npm_config_audit = 'false'
 
-var npm_config_cache = path.resolve(__dirname, 'npm_cache_' + testId)
-ourenv.npm_config_cache = exports.npm_config_cache = npm_config_cache
+ourenv.npm_config_unsafe_perm = 'true'
+ourenv.npm_config_cache = commonCache
 ourenv.npm_config_userconfig = exports.npm_config_userconfig = configCommon.userconfig
 ourenv.npm_config_globalconfig = exports.npm_config_globalconfig = configCommon.globalconfig
 ourenv.npm_config_global_style = 'false'
@@ -88,7 +131,10 @@ exports.npm = function (cmd, opts, cb) {
   opts.env = opts.env || process.env
   if (opts.env._storage) opts.env = Object.assign({}, opts.env._storage)
   if (!opts.env.npm_config_cache) {
-    opts.env.npm_config_cache = npm_config_cache
+    opts.env.npm_config_cache = commonCache
+  }
+  if (!opts.env.npm_config_unsafe_perm) {
+    opts.env.npm_config_unsafe_perm = 'true'
   }
   if (!opts.env.npm_config_send_metrics) {
     opts.env.npm_config_send_metrics = 'false'
